refactor: move theme atom to common package
