+++ .gitignore
@@ -18,3 +18,5 @@
 
 # Node Modules
 node_modules/
+
+*.tsbuildinfo
