+++ biome.jsonc
@@ -8,9 +8,9 @@
   "files": {
     "ignoreUnknown": false,
     "includes": [
-      "**/packages/frontend/src/**/*.ts",
-      "**/packages/frontend/src/**/*.tsx",
-      "**/packages/frontend/src/**/*.astro",
+      "**/packages/*/src/**/*.ts",
+      "**/packages/*/src/**/*.tsx",
+      "**/packages/*/src/**/*.astro",
       "**/*/biome.jsonc",
       "**/*/tsconfig.json",
       "**/*/astro.config.ts"
