+++ packages/common/package.json
@@ -1,13 +1,17 @@
 {
-  "name": "@furtherland/post-components",
+  "name": "@furtherland/common",
   "version": "0.0.0",
   "private": true,
   "type": "module",
+  "dependencies": {
+    "nanostores": "^1.5.2"
+  },
   "devDependencies": {
     "@biomejs/biome": "^2.5.11",
     "@types/node": "^26.4.1",
     "@types/react": "^19.2.18",
     "@types/react-dom": "^19.2.5",
+    "astro": "^7.2.10",
     "typescript": "6"
   }
 }
